move tsbuildinfo

+++ setup_ts.py
@@ -191,7 +191,7 @@ def copy_nni_node(version):
 
     # copytree(..., dirs_exist_ok=True) is not supported by Python 3.6
     for path in Path('ts/nni_manager/dist').iterdir():
-        if path.is_file():
+        if path.is_file() and path.name != 'nni_manager.tsbuildinfo':
             shutil.copyfile(path, Path('nni_node', path.name))
         else:
             shutil.copytree(path, Path('nni_node', path.name))
@@ -240,7 +240,6 @@ def _print(*args, color='cyan'):
 
 
 generated_files = [
-    'ts/nni_manager/.tsbuildinfo',
     'ts/nni_manager/dist',
     'ts/nni_manager/node_modules',
     'ts/webui/build',
